Rename DataSourceFactory to HikariDataSourceFactory; add base DataSourceFactory interface
