+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc-apps/src/main/java/gov/cms/bfd/pipeline/rda/grpc/apps/LoadRdaJsonApp.java
@@ -17,9 +17,9 @@ import gov.cms.bfd.pipeline.sharedutils.PipelineJob;
 import gov.cms.bfd.pipeline.sharedutils.s3.S3ClientConfig;
 import gov.cms.bfd.sharedutils.config.AwsClientConfig;
 import gov.cms.bfd.sharedutils.config.ConfigLoader;
-import gov.cms.bfd.sharedutils.database.DataSourceFactory;
 import gov.cms.bfd.sharedutils.database.DatabaseOptions;
 import gov.cms.bfd.sharedutils.database.DatabaseSchemaManager;
+import gov.cms.bfd.sharedutils.database.DefaultHikariDataSourceFactory;
 import gov.cms.bfd.sharedutils.database.HikariDataSourceFactory;
 import gov.cms.bfd.sharedutils.database.RdsDataSourceFactory;
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
@@ -97,13 +97,13 @@ public class LoadRdaJsonApp {
                 final RdaLoadOptions jobConfig = config.createRdaLoadOptions(port);
                 final DatabaseOptions databaseConfig = config.createDatabaseOptions();
                 final AwsClientConfig awsClientConfig = config.createAwsClientConfig();
-                final DataSourceFactory dataSourceFactory =
+                final HikariDataSourceFactory dataSourceFactory =
                     databaseConfig.getAuthenticationType() == DatabaseOptions.AuthenticationType.RDS
                         ? RdsDataSourceFactory.builder()
                             .awsClientConfig(awsClientConfig)
                             .databaseOptions(databaseConfig)
                             .build()
-                        : new HikariDataSourceFactory(databaseConfig);
+                        : new DefaultHikariDataSourceFactory(databaseConfig);
                 final HikariDataSource pooledDataSource =
                     PipelineApplicationState.createPooledDataSource(dataSourceFactory, metrics);
                 if (config.runSchemaMigration) {
