+++ apps/bfd-shared-utils/src/main/java/gov/cms/bfd/sharedutils/database/DataSourceFactory.java
@@ -1,13 +1,14 @@
 package gov.cms.bfd.sharedutils.database;
 
-import com.zaxxer.hikari.HikariDataSource;
+import javax.sql.DataSource;
 
-/** Implementations of this interface create {@link HikariDataSource} instances on demand. */
+/** Implementations of this interface create {@link DataSource} instances on demand. */
 public interface DataSourceFactory {
+
   /**
-   * Create a properly configured {@link HikariDataSource} instance.
+   * Create a properly configured {@link DataSource} instance.
    *
    * @return the instance
    */
-  HikariDataSource createDataSource();
+  DataSource createDataSource();
 }
