+++ apps/bfd-shared-utils/src/main/java/gov/cms/bfd/sharedutils/database/HikariDataSourceFactory.java
@@ -1,34 +1,13 @@
 package gov.cms.bfd.sharedutils.database;
 
 import com.zaxxer.hikari.HikariDataSource;
-import lombok.AllArgsConstructor;
 
-/**
- * Simple implementation of {@link DataSourceFactory} that creates a {@link HikariDataSource}
- * instance using settings defined in a {@link DatabaseOptions} instance.
- */
-@AllArgsConstructor
-public class HikariDataSourceFactory implements DataSourceFactory {
-  /** Used to configure constructed instances. */
-  private final DatabaseOptions dbOptions;
-
-  @Override
-  public HikariDataSource createDataSource() {
-    HikariDataSource pooledDataSource = new HikariDataSource();
-    configureDataSource(pooledDataSource);
-    return pooledDataSource;
-  }
-
+/** Implementations of this interface create {@link HikariDataSource} instances on demand. */
+public interface HikariDataSourceFactory extends DataSourceFactory {
   /**
-   * Applies settings from {@link HikariDataSourceFactory#dbOptions} to the provided data source.
+   * Create a properly configured {@link HikariDataSource} instance.
    *
-   * @param dataSource data source to configure
+   * @return the instance
    */
-  protected void configureDataSource(HikariDataSource dataSource) {
-    dataSource.setJdbcUrl(dbOptions.getDatabaseUrl());
-    dataSource.setUsername(dbOptions.getDatabaseUsername());
-    dataSource.setPassword(dbOptions.getDatabasePassword());
-    dataSource.setMaximumPoolSize(Math.max(2, dbOptions.getMaxPoolSize()));
-    dataSource.setRegisterMbeans(true);
-  }
+  HikariDataSource createDataSource();
 }
